add jira-helpdesk on every page
